Revamp vim configuration.
Going back to basics with my vim configuration as my previous
configuration was an absolute clusterfuck of options and code
and I didn't even understand half of it. So now I'm going back
to basics and I will expand the vim configuration as I find more
settings that I think will benefit me. Also have to setup vim-plug
at some point but that'll be later.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>